refactor: move complexity section contracts to engine
